+++ services/openvpn/discovery/service_poposal_test.go
@@ -22,8 +22,8 @@ import (
 	"testing"
 
 	nats_discovery "github.com/mysteriumnetwork/node/communication/nats/discovery"
-	"github.com/mysteriumnetwork/node/openvpn"
 	dto_discovery "github.com/mysteriumnetwork/node/service_discovery/dto"
+	"github.com/mysteriumnetwork/node/services/openvpn"
 	dto_openvpn "github.com/mysteriumnetwork/node/services/openvpn/discovery/dto"
 	"github.com/stretchr/testify/assert"
 )
